Use the new product page everywhere
